refactor: 使用状況入力エリアを切り出す
