+++ frontend_challenge_2/challenges/toru_shimizu/coding-challenge/src/components/features/usage-condition.module.scss
@@ -1,0 +1,4 @@
+.usageCondition {
+  display: grid;
+  gap: 20px;
+}
